Updates
Adding some requirements

Sign In button must not be enabled until username and password entries are populated

If the account doesn’t exist, an error message should show up on the screen, specifying that the account doesn’t exist.

If the password doesn’t match, an error message should show up on the screen specifying that the password is incorrect.

If the user account exists and the password matches, a success message should show up on the screen (message up to the candidate).